build: Fix module loading order to avoid module path lost
When using cross-compile environments we usually provide a toolchain
file and it is good practice to ensure the toolchain CMake is used for
build (thus a explicit and a pre-defined CMAKE_MODULE_PATH should be
provided).

The CMakeDetermineSystem.cmake module is responsable to load the
toolchain provided settings thus overriding the settings of the build
system and thus overriding the CMAKE_MODULE_PATH value.

To fix it, the system detection needs to be done first and after that
regular modules can be used as usual making it cross-compile friendly.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>